BML full-cycle code auto-commit (iteration 1 of 1)
# ─── Retro→Code Linking Tags (optional but recommended) ───
# Add tags to link commits to retrospectives and track WSJF metrics:
#
# [QW-ID:xxx] [RETRO-ID:yyy] [WSJF:x.y] [CoD:$zzz] [Repo:agentic-flow]
#
# Example:
# [QW-ID:NOW-1] [RETRO-ID:2025-11-13] [WSJF:10.0] [CoD:$500/day] [Repo:agentic-flow]
#
# Quick reference:
# - QW-ID: Quick wins item ID (NOW-1, NEXT-4, etc.)
# - RETRO-ID: Retrospective date or ID
# - WSJF: Weighted Shortest Job First score
# - CoD: Cost of Delay estimate
# - Repo: Repository name for cross-repo tracking

+++ scripts/emit_metrics.py
@@ -123,8 +123,8 @@ def main():
     )
     parser.add_argument(
         "--risk-score",
-        type=int,
-        default=0,
+        type=float,
+        default=0.0,
         help="Current system risk score"
     )
     parser.add_argument(
